requirements.txt
transformers==4.37.2
Pillow==10.4.0
numpy==1.26.4
tqdm==4.66.4
scikit-learn==1.3.2
accelerate==0.21.0
bitsandbytes==0.43.1
einops==0.6.1
einops-exts==0.0.4
timm==0.6.13
sentencepiece==0.1.99
shortuuid==1.0.13
termcolor==2.4.0
jsonlines==4.0.0
openai-clip
scipy==1.10.1
hdbscan==0.8.38.post1
umap-learn==0.5.6 
sentence-transformers==3.0.1
bertopic==0.16.3
requests==2.32.3